Stability, HTTP headers & improved console logging
Form inspector now shows the HTTP headers of the request in the response shown to the user. Also, logging to the console is improved (looks prettier). 

As for stability, numerous bugs were fixed in detecting and handling empty requests, and for such requests we now return an error to the client (assuming that responding to empty requests has been disabled on line 13). 

I have also removed the XFF IP from fi_requestIP in the response, as this information is now located in the headers section.